Added missing requirement and bumped django-rq min version.
After support for cron based scheduling was added a dependancy for `croniter` was added to `models.py` that was not reflected in `setup.py` or `README.md`.
Also, during testing / usage an error was noticed that I tracked down to BaseCommand.option_list deprecated in django 1.10 while django-rq 0.9.3 was attempting to use it. (The error appeared when running `manage.py`.)
Lastly, I believe rq-scheduler was mis-named as rq_scheduler.